refactor(Results): add GuessGradient
