+++ components/Results.tsx
@@ -5,6 +5,7 @@ import ResultsProps from '../interfaces/ResultsProps';
 import ColorDisplayer from './Game/ColorDisplayer';
 import Popup from './Popup';
 import Diagram from './Results/Diagram';
+import GuessGradient from './Results/GuessGradient';
 
 const calculateTimeLeft = () => {
 	const now = new Date();
@@ -27,8 +28,6 @@ const calculateTimeLeft = () => {
 const Results = ({ attempts, close, color, displayed, ended, guesses }: ResultsProps) => {
 
 	const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
-	const canvasRef = useRef<HTMLCanvasElement>(null);
-	const gradientCopyRef = useRef<HTMLSpanElement>(null);
 	const shareIconRef = useRef<HTMLDivElement>(null);
 	const checkIconRef = useRef<HTMLDivElement>(null);
 
@@ -39,22 +38,6 @@ const Results = ({ attempts, close, color, displayed, ended, guesses }: ResultsP
 		return () => clearTimeout(timer);
 	});
 
-	useEffect(() => {
-		if (!guesses?.length)
-			return;
-		if (canvasRef.current) {
-			const ctx = canvasRef.current.getContext('2d');
-			if (!ctx)
-				return;
-			const gradient = ctx.createLinearGradient(0, 0, 0, canvasRef.current.height);
-			for (const i in guesses) {
-				const guess = guesses[i];
-				gradient.addColorStop(parseInt(i) / guesses.length, `rgba(${guess[0]}, ${guess[1]}, ${guess[2]})`);
-			}
-			ctx.fillStyle = gradient;
-			ctx.fillRect(0, 0, canvasRef.current.width, canvasRef.current.height);
-		}
-	}, [guesses]);
 
 	let currentStreak = 0;
 	attempts.reverse();
@@ -95,23 +78,6 @@ const Results = ({ attempts, close, color, displayed, ended, guesses }: ResultsP
 			shareIconRef.current!.classList.remove('scale-0');
 			checkIconRef.current!.classList.add('scale-0');
 		}, 2000);
-	}
-
-	const copyGradient = () => {
-		const canvas = canvasRef.current;
-		if (!canvas)
-			return;
-		canvas.toBlob(blob => {
-			if (blob) {
-				navigator.clipboard.write([new ClipboardItem({ 'image/png': blob })]);
-				if (!gradientCopyRef.current)
-					return;
-				gradientCopyRef.current.innerText = 'Copied!';
-				setTimeout(() => {
-					gradientCopyRef.current!.innerText = 'Click to copy';
-				}, 2000);
-			}
-		});
 	}
 
 	return (
@@ -195,25 +161,7 @@ const Results = ({ attempts, close, color, displayed, ended, guesses }: ResultsP
 				<>
 					<div className='my-4 text-lg md:text-2xl text-center font-title'>Guess Gradient</div>
 					<div className='flex justify-center'>
-						<div
-							className='w-11/12 rounded overflow-hidden cursor-pointer'
-							onClick={copyGradient}
-						>
-							<canvas
-								ref={canvasRef}
-								className='object-fit w-full'
-								width={1920}
-								height={1080}
-							/>
-							<div
-								className='absolute top-0 left-0 w-full h-full flex justify-center items-center
-								bg-black/40 font-title text-xl md:text-4xl
-								opacity-0 hover:opacity-100 transition-opacity duration-100'
-							>
-								<FaRegClipboard className='mr-2' />
-								<span ref={gradientCopyRef}>Click to copy</span>
-							</div>
-						</div>
+						<GuessGradient guesses={guesses!} />
 					</div>
 				</>
