+++ interfaces/Results/GuessGradientProps.ts
@@ -1,0 +1,7 @@
+/**
+ * Props for `GuessGradient` component.
+ */
+interface GuessGradientProps {
+	guesses: number[][];
+}
+export default GuessGradientProps;
